Se aplico Encuestas del doc arreglar

+++ app/Http/Controllers/PreguntaController.php
@@ -3,13 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Encuesta;
-use App\Pregunta;
-use Validator;
 use App\Http\Traits\Utilidades;
+use App\Pregunta;
+use App\Realizada;
 use App\User;
 use DB;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Validator;
 
 class PreguntaController extends Controller
 {
@@ -76,12 +77,15 @@ class PreguntaController extends Controller
         //     $validator->errors()->add('Repetido', 'enunciado ya existe');            
         // }
 
-
-
-
-
         $encuesta = Encuesta::findOrFail($id);
         $preguntas = Pregunta::where('encuesta_id',$id)->get();
+
+        if (Realizada::where('encuesta_id', $id)->get()->isNotEmpty()) {
+            $request->session()->flash(
+                'message',
+                'Imposible agregar pregunta, esta encuesta ya fue completada por alguien');
+            return view('pregunta.create',['encuesta'=>$encuesta, 'preguntas'=>$preguntas]);
+        }
 
         if($validator->fails() ){
             return view('pregunta.create',['encuesta'=>$encuesta, 'preguntas'=>$preguntas])->withErrors($validator,'enunciado');
@@ -144,8 +148,15 @@ class PreguntaController extends Controller
         $pregunta = Pregunta::findOrFail($idPreg);
         $idEncuesta = $pregunta->encuesta()->get()[0]->id;
         $pregunta->enunciado = $request->get('enunciado');
-        $pregunta->save();
-        $request->session()->flash('message', 'Pregunta actualizada exitosmente!');
+
+        if (Realizada::where('encuesta_id', $idEncuesta)->get()->isNotEmpty()) {
+            $request->session()->flash(
+                'message',
+                'Imposible modificar pregunta, esta encuesta ya fue completada por alguien');
+        } else {
+            $pregunta->save();
+            $request->session()->flash('message', 'Pregunta actualizada exitosmente!');
+        }
 
         return $this->create($idEncuesta);
 
@@ -162,7 +173,15 @@ class PreguntaController extends Controller
         $this->authorize('es_admin', User::class);
 
         $pregunta = Pregunta::findOrFail($idPregunta);
-        $idEncuesta = $pregunta->encuesta()->get()[0]->id;  
+        $idEncuesta = $pregunta->encuesta()->get()[0]->id;
+
+        if (Realizada::where('encuesta_id', $idEncuesta)->get()->isNotEmpty()) {
+            $request->session()->flash(
+                'message',
+                'Imposible quitar pregunta, esta encuesta ya fue completada por alguien');
+            return $this->create($idEncuesta);
+        }
+
         $pregunta->delete();
 
         //$table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
